Add files via upload
The "Flight Status Analysis" project aims to develop a comprehensive and visually engaging 
Power BI dashboard to analyze and monitor the status of flights. This dashboard will leverage 
Power BI's robust features to provide stakeholders, including airline operators, airport 
authorities, and travelers, with valuable insights into the performance and status of flights.
The total number of flights (incoming and departing flights), the total number of flight delays 
(arrivals and departures), ground processing time, and the passenger feedback section can all 
be included in the summary page of this project’s dashboard. In case of flight changes, 
emergencies, or delays, this analysis could assist airport management authorities in making 
timely data-driven choices. It helps in analyzing the flight status(on time, delayed, cancelled) 
month wise and airline wise. Also have a look at average delay time for that duration. Air 
traffic analysis is performed which gives insights as when is the best time to book a flight and 
what is the peak hour when the maximum flights are available.